requirements.txt
selene==2.0.0rc4
selenium==4.25.0
pytest==8.3.3

allure-pytest==2.13.5
allure-python-commons==2.13.5

python-dotenv==1.0.1
requests==2.32.3
